create drawing, objsettings and return context in LJV
